Revert "Merge branch 'features/docs' into features/review"
This reverts commit 4c11870bd461237a30a291e7af863de0ae791373, reversing
changes made to 5574a18d7a501a03d47abffc5a5e1c3b4e91f2fc.
