+++ src/main/java/org/guzzing/studayserver/domain/auth/config/Constants.java
@@ -4,13 +4,6 @@ public class Constants {
 
     public static final String[] permitAllArray = new String[]{
             "/",
-            "/auth/**",
-            "/error",
-            "/docs/**",
-            "/favicon.ico",
-            "/swagger-ui/**",
-            "/swagger-resources/**",
-            "/v3/api-docs/**",
-            "/src/main/resources/static/swagger.json"
+            "/auth/**"
     };
 }
