+++ src/test/java/org/guzzing/studayserver/domain/like/controller/LikeRestControllerTest.java
@@ -1,13 +1,13 @@
 package org.guzzing.studayserver.domain.like.controller;
 
-import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
-import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
 import static org.guzzing.studayserver.testutil.fixture.TestConfig.AUTHORIZATION_HEADER;
 import static org.guzzing.studayserver.testutil.fixture.TestConfig.BEARER;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
@@ -18,13 +18,13 @@ import static org.springframework.restdocs.payload.JsonFieldType.ARRAY;
 import static org.springframework.restdocs.payload.JsonFieldType.NUMBER;
 import static org.springframework.restdocs.payload.JsonFieldType.STRING;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
-import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
+import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.guzzing.studayserver.domain.academy.service.AcademyAccessService;
 import org.guzzing.studayserver.domain.like.controller.dto.request.LikePostRequest;
@@ -52,8 +52,6 @@ import org.springframework.transaction.annotation.Transactional;
 @SpringBootTest
 @Transactional
 class LikeRestControllerTest {
-
-    private static final String TAG = "좋아요 API";
 
     @Autowired
     private MockMvc mockMvc;
@@ -107,21 +105,16 @@ class LikeRestControllerTest {
                 .andDo(document("post-like",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint()),
-                        resource(ResourceSnippetParameters.builder()
-                                .tag(TAG)
-                                .summary("좋아요 등록")
-                                .requestHeaders(
-                                        headerWithName("Authorization").description("JWT 토큰 (Bearer)")
-                                )
-                                .requestFields(
-                                        fieldWithPath("academyId").type(NUMBER).description("학원 아이디")
-                                )
-                                .responseFields(
-                                        fieldWithPath("likeId").type(NUMBER).description("좋아요 아이디"),
-                                        fieldWithPath("memberId").type(NUMBER).description("학원 아이디"),
-                                        fieldWithPath("academyId").type(NUMBER).description("학원 아이디")
-                                )
-                                .build()
+                        requestHeaders(
+                                headerWithName("Authorization").description("JWT 토큰 (Bearer)")
+                        ),
+                        requestFields(
+                                fieldWithPath("academyId").type(NUMBER).description("학원 아이디")
+                        ),
+                        responseFields(
+                                fieldWithPath("likeId").type(NUMBER).description("좋아요 아이디"),
+                                fieldWithPath("memberId").type(NUMBER).description("학원 아이디"),
+                                fieldWithPath("academyId").type(NUMBER).description("학원 아이디")
                         )
                 ));
     }
@@ -146,13 +139,8 @@ class LikeRestControllerTest {
                 .andDo(document("delete-like",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint()),
-                        resource(ResourceSnippetParameters.builder()
-                                .tag(TAG)
-                                .summary("좋아요 제거")
-                                .pathParameters(
-                                        parameterWithName("likeId").description("좋아요 아이디")
-                                )
-                                .build()
+                        requestHeaders(
+                                headerWithName("Authorization").description("JWT 토큰 (Bearer)")
                         )
                 ));
     }
@@ -181,18 +169,12 @@ class LikeRestControllerTest {
                 .andDo(document("get-like",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint()),
-                        resource(ResourceSnippetParameters.builder()
-                                .tag(TAG)
-                                .summary("나의 좋아요 조회")
-                                .responseFields(
-                                        fieldWithPath("likeAcademyInfos").type(ARRAY).description("좋아요한 학원 비용 목록"),
-                                        fieldWithPath("likeAcademyInfos[].academyId").type(NUMBER)
-                                                .description("좋아요한 학원 아이디"),
-                                        fieldWithPath("likeAcademyInfos[].academyName").type(STRING).description("학원명"),
-                                        fieldWithPath("likeAcademyInfos[].expectedFee").description("예상 교육비"),
-                                        fieldWithPath("totalFee").type(NUMBER).description("총 비용")
-                                )
-                                .build()
+                        responseFields(
+                                fieldWithPath("likeAcademyInfos").type(ARRAY).description("좋아요한 학원 비용 목록"),
+                                fieldWithPath("likeAcademyInfos[].academyId").type(NUMBER).description("좋아요한 학원 아이디"),
+                                fieldWithPath("likeAcademyInfos[].academyName").type(STRING).description("학원명"),
+                                fieldWithPath("likeAcademyInfos[].expectedFee").description("예상 교육비"),
+                                fieldWithPath("totalFee").type(NUMBER).description("총 비용")
                         )
                 ));
     }
