+++ src/test/java/org/guzzing/studayserver/domain/review/controller/ReviewRestControllerTest.java
@@ -1,12 +1,11 @@
 package org.guzzing.studayserver.domain.review.controller;
 
-import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
-import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
-import static com.epages.restdocs.apispec.Schema.schema;
 import static org.guzzing.studayserver.testutil.fixture.TestConfig.AUTHORIZATION_HEADER;
 import static org.guzzing.studayserver.testutil.fixture.TestConfig.BEARER;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
@@ -15,13 +14,15 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.payload.JsonFieldType.BOOLEAN;
 import static org.springframework.restdocs.payload.JsonFieldType.NUMBER;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
+import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.guzzing.studayserver.domain.academy.listener.NewReviewListener;
 import org.guzzing.studayserver.domain.academy.service.AcademyAccessService;
@@ -34,24 +35,19 @@ import org.guzzing.studayserver.testutil.fixture.TestConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.transaction.annotation.Transactional;
 
 @AutoConfigureMockMvc
 @AutoConfigureRestDocs
-@ExtendWith(RestDocumentationExtension.class)
 @SpringBootTest
 class ReviewRestControllerTest {
-
-    private static final String TAG = "리뷰 API";
 
     @Autowired
     private MockMvc mockMvc;
@@ -100,30 +96,21 @@ class ReviewRestControllerTest {
                 .andDo(document("post-review",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint()),
-                        resource(ResourceSnippetParameters.builder()
-                                .tag(TAG)
-                                .summary("리뷰 등록")
-                                .requestHeaders(
-                                        headerWithName("Authorization").description("JWT 토큰 (Bearer)")
-                                )
-                                .requestFields(
-                                        fieldWithPath("academyId").type(NUMBER).description("학원 아이디"),
-                                        fieldWithPath("kindness").type(BOOLEAN).description("친절해요 리뷰 선택 여부"),
-                                        fieldWithPath("cheapFee").type(BOOLEAN).description("수강료가 싸요 리뷰 선택 여부"),
-                                        fieldWithPath("goodFacility").type(BOOLEAN).description("시설이 좋아요 리뷰 선택 여부"),
-                                        fieldWithPath("goodManagement").type(BOOLEAN).description("관리가 좋아요 리뷰 선택 여부"),
-                                        fieldWithPath("lovelyTeaching").type(BOOLEAN)
-                                                .description("가르침이 사랑스러워요 리뷰 선택 여부"),
-                                        fieldWithPath("shuttleAvailability").type(BOOLEAN)
-                                                .description("셔틀을 운행해요 리뷰 선택 여부")
-                                )
-                                .responseFields(
-                                        fieldWithPath("reviewId").type(NUMBER).description("리뷰 아이디"),
-                                        fieldWithPath("academyId").type(NUMBER).description("학원 아이디")
-                                )
-                                .requestSchema(schema("요청?"))
-                                .responseSchema(schema("응답?"))
-                                .build()
+                        requestHeaders(
+                                headerWithName("Authorization").description("JWT 토큰 (Bearer)")
+                        ),
+                        requestFields(
+                                fieldWithPath("academyId").type(NUMBER).description("학원 아이디"),
+                                fieldWithPath("kindness").type(BOOLEAN).description("친절해요 리뷰 선택 여부"),
+                                fieldWithPath("cheapFee").type(BOOLEAN).description("수강료가 싸요 리뷰 선택 여부"),
+                                fieldWithPath("goodFacility").type(BOOLEAN).description("시설이 좋아요 리뷰 선택 여부"),
+                                fieldWithPath("goodManagement").type(BOOLEAN).description("관리가 좋아요 리뷰 선택 여부"),
+                                fieldWithPath("lovelyTeaching").type(BOOLEAN).description("가르침이 사랑스러워요 리뷰 선택 여부"),
+                                fieldWithPath("shuttleAvailability").type(BOOLEAN).description("셔틀을 운행해요 리뷰 선택 여부")
+                        ),
+                        responseFields(
+                                fieldWithPath("reviewId").type(NUMBER).description("리뷰 아이디"),
+                                fieldWithPath("academyId").type(NUMBER).description("학원 아이디")
                         )
                 ));
     }
@@ -152,20 +139,15 @@ class ReviewRestControllerTest {
                 .andDo(document("get-reviewable",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint()),
-                        resource(ResourceSnippetParameters.builder()
-                                .tag(TAG)
-                                .summary("리뷰 가능 여부 확인")
-                                .requestHeaders(
-                                        headerWithName("Authorization").description("JWT 토큰 (Bearer)")
-                                )
-                                .queryParameters(
-                                        parameterWithName("academyId").description("학원 아이디")
-                                )
-                                .responseFields(
-                                        fieldWithPath("academyId").type(NUMBER).description("학원 아이디"),
-                                        fieldWithPath("reviewable").type(BOOLEAN).description("리뷰 등록 가능 여부")
-                                )
-                                .build()
+                        requestHeaders(
+                                headerWithName("Authorization").description("JWT 토큰 (Bearer)")
+                        ),
+                        queryParameters(
+                                parameterWithName("academyId").description("학원 아이디")
+                        ),
+                        responseFields(
+                                fieldWithPath("academyId").type(NUMBER).description("학원 아이디"),
+                                fieldWithPath("reviewable").type(BOOLEAN).description("리뷰 등록 가능 여부")
                         )
                 ));
     }
